Set up new finnts forecast project

Creates list object of information helpful in logging information about your entire forecast project.

set_project_info(
  project_name = "finn_project",
  path = NULL,
  combo_variables,
  target_variable,
  date_type,
  fiscal_year_start = 1,
  weekly_to_daily = TRUE,
  storage_object = NULL,
  data_output = "csv",
  object_output = "rds",
  overwrite = FALSE
)

Arguments

project_name

Name used to group similar runs under a single project name.

path

String showing what file path the outputs should be written to. Default of NULL will write the outputs to a temporary directory within R, which will delete itself after the R session closes.

combo_variables

Character vector of variables to combine into a combo variable.

target_variable

Character string of the target variable to forecast.

date_type

Character string of the type of date variable

fiscal_year_start

Numeric value of the month that the fiscal year starts in.

weekly_to_daily

Logical value of whether to convert weekly data to daily data. Default of FALSE will not convert weekly data to daily data.

storage_object

Used to store outputs during the project to other storage services in Azure. Could be a storage container object from the 'AzureStor' package to connect to ADLS blob storage or a OneDrive/SharePoint object from the 'Microsoft365R' package to connect to a OneDrive folder or SharePoint site. Default of NULL will save outputs to the local file system.

data_output

String value describing the file type for data outputs. Default will write data frame outputs as csv files. The other option of 'parquet' will instead write parquet files.

object_output

String value describing the file type for object outputs. Default will write object outputs like trained models as rds files. The other option of 'qs' will instead serialize R objects as qs files by using the 'qs' package.

overwrite

Logical value of whether to overwrite existing project

Value

A list of project information

Examples

if (FALSE) { # \dontrun{
project_info <- set_project_info(
  project_name = "test_project",
  combo_variables = c("Store", "Product"),
  target_variable = "Sales",
  date_type = "month"
)
} # }