Set up finnts submission

Creates list object of information helpful in logging information about your run.

set_run_info(
  experiment_name = "finn_fcst",
  run_name = "finn_fcst",
  storage_object = NULL,
  path = NULL,
  data_output = "csv",
  object_output = "rds",
  add_unique_id = TRUE
)

Arguments

experiment_name

Name used to group similar runs under a single experiment name.

run_name

Name to distinguish one run of Finn from another. The current time in UTC is appended to the run name to ensure a unique run name is created.

storage_object

Used to store outputs during a run to other storage services in Azure. Could be a storage container object from the 'AzureStor' package to connect to ADLS blob storage or a OneDrive/SharePoint object from the 'Microsoft365R' package to connect to a OneDrive folder or SharePoint site. Default of NULL will save outputs to the local file system.

path

String showing what file path the outputs should be written to. Default of NULL will write the outputs to a temporary directory within R, which will delete itself after the R session closes.

data_output

String value describing the file type for data outputs. Default will write data frame outputs as csv files. The other option of 'parquet' will instead write parquet files.

object_output

String value describing the file type for object outputs. Default will write object outputs like trained models as rds files. The other option of 'qs' will instead serialize R objects as qs files by using the 'qs' package.

add_unique_id

Add a unique id to end of run_name based on submission time. Set to FALSE to supply your own unique run name, which is helpful in multistage ML pipelines.

Value

A list of run information

Examples

# \donttest{
run_info <- set_run_info(
  experiment_name = "test_exp",
  run_name = "test_run_1"
)
#> Finn Submission Info
#> • Experiment Name: test_exp
#> • Run Name: test_run_1-20230717T200541Z
#> 
# }