Bridge prediction Function for SVM-POLY Multistep Horizon Models

Bridge prediction Function for SVM-POLY Multistep Horizon Models

svm_poly_multistep_predict_impl(object, new_data, ...)

Arguments

object

model object

new_data

input data to predict

...

Additional parsnip-related options, depending on the value of type. Arguments to the underlying model's prediction function cannot be passed here (use the opts argument instead). Possible arguments are:

interval: for type equal to "survival" or "quantile", should interval estimates be added, if available? Options are "none" and "confidence".

level: for type equal to "conf_int", "pred_int", or "survival", this is the parameter for the tail area of the intervals (e.g. confidence level for confidence intervals). Default value is 0.95.

std_error: for type equal to "conf_int" or "pred_int", add the standard error of fit or prediction (on the scale of the linear predictors). Default value is FALSE.

quantile: for type equal to quantile, the quantiles of the distribution. Default is (1:9)/10.

eval_time: for type equal to "survival" or "hazard", the time points at which the survival probability or hazard is estimated.

Value

predictions